+++ app/src/main/java/ru/bookshop/ui/screens/account_edit/models/AccountEditEvent.kt
@@ -1,4 +1,7 @@
 package ru.bookshop.ui.screens.account_edit.models
 
-class AccountEditEvent {
+import ru.bookshop.data.models.AccountDTO
+
+sealed class AccountEditEvent {
+    data class OnDoneClick(val info: AccountDTO): AccountEditEvent()
 }
