+++ app/src/main/java/ru/bookshop/ui/screens/account_edit/presentation/AccountEditViewModel.kt
@@ -1,6 +1,9 @@
 package ru.bookshop.ui.screens.account_edit.presentation
 
+import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
 import ru.bookshop.data.models.AccountDTO
 import ru.bookshop.ui.common.models.BaseViewModel
 import ru.bookshop.ui.screens.account_edit.models.AccountEditEvent
@@ -11,6 +14,22 @@ import javax.inject.Inject
 class AccountEditViewModel @Inject constructor() :
     BaseViewModel<AccountEditViewState, AccountEditEvent>(AccountEditViewState()) {
 
-    private val accountInfo = AccountDTO()
-
+    override fun obtainEvent(viewEvent: AccountEditEvent) {
+        when (viewEvent) {
+            is AccountEditEvent.OnDoneClick -> {
+                viewModelScope.launch {
+                    viewState.update {
+                        it.copy(
+                            accountInfo = AccountDTO(
+                                image = viewEvent.info.image,
+                                name = viewEvent.info.name,
+                                job = viewEvent.info.job,
+                                resumeUrl = viewEvent.info.resumeUrl,
+                            )
+                        )
+                    }
+                }
+            }
+        }
+    }
 }
