--- app/src/main/java/ru/bookshop/ui/screens/account_edit/ui/views/AccountEditTitle.kt
@@ -1,64 +1,0 @@
-package ru.bookshop.ui.screens.account_edit.ui.views
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import ru.bookshop.R
-import ru.bookshop.ui.theme.BookshopTheme
-
-@Composable
-fun AccountEditTitle(
-    onDoneClick: () -> Unit,
-    onBackClick: () -> Unit,
-) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(bottom = 16.dp),
-    ) {
-        IconButton(onBackClick) {
-            Icon(
-                painter = painterResource(R.drawable.ic_arrow_back),
-                contentDescription = "go back"
-            )
-        }
-
-        Text(
-            text = "Редактирование",
-            style = MaterialTheme.typography.titleLarge,
-            color = Color.Black,
-        )
-
-        IconButton(onDoneClick) {
-            Icon(
-                painter = painterResource(R.drawable.ic_done),
-                contentDescription = "done profile"
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun PreviewAccountTitle() {
-    BookshopTheme {
-        AccountEditTitle(
-            onBackClick = {},
-            onDoneClick = {},
-        )
-    }
-}
